@@ -5,7 +5,7 @@
 
     <groupId>com.example</groupId>
     <artifactId>devops</artifactId>
-    <version>0.0.9-RELEASE</version>
+    <version>2020.1.0</version>
     <packaging>jar</packaging>
 
     <name>devops</name>
@@ -14,15 +14,15 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.3.1.RELEASE</version>
+        <version>2.4.1</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>14</java.version>
-        <jacoco.version>0.8.5</jacoco.version>
+        <java.version>15</java.version>
+        <jacoco.version>0.8.6</jacoco.version>
         <asciidoctor-maven-plugin.version>1.6.0</asciidoctor-maven-plugin.version>
 
         <!-- Sonar properties -->
