@@ -1,10 +1,5 @@
 /*----------------------------------------------------------------------------*/
 /* Source File:   HOMECONTROLLERINTEGRATIONTEST.JAVA                          */
-/* Description:   Test REST Api for Home end-points.                          */
-/* Author:        Carlos Adolfo Ortiz Quirós (COQ)                            */
-/* Date:          Mar.01/2019                                                 */
-/* Last Modified: Jul.24/2020                                                 */
-/* Version:       1.1                                                         */
 /* Copyright (c), 2019, 2020 CSoftZ                                           */
 /*----------------------------------------------------------------------------*/
 /*-----------------------------------------------------------------------------
@@ -23,13 +18,13 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.http.MediaType;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
 
 import com.example.devops.api.controller.HomeController;
@@ -40,13 +35,13 @@ import com.example.devops.api.controller.HomeController;
  * <a href="https://spring.io/guides/gs/testing-restdocs/#initial">Spring Rest Docs Guide</a>
  *
  * @author Carlos Adolfo Ortiz Quirós (COQ)
- * @version 1.1, Jul.24/2020
+ * @version 1.1, Dec.12/2020
  * @since 11 (JDK), Mar.01/2019
  */
-@RunWith(SpringRunner.class)
+@ExtendWith(SpringExtension.class)
 @WebMvcTest(HomeController.class)
 @AutoConfigureRestDocs(uriScheme = "http", uriHost = "api.example.com")
-public class HomeControllerIntegrationTest {
+class HomeControllerIntegrationTest {
     private static final String HOME_URL = "/";
     private static final String ASCIIDOC_SNIPPET_HOME = "home";
     private static final String ASCIIDOC_FIELD_DESC_MESSAGE = "message";
@@ -63,7 +58,7 @@ public class HomeControllerIntegrationTest {
      * @throws Exception Reports some error.
      */
     @Test
-    public void shouldReturnDefaultMessage() throws Exception {
+    void shouldReturnDefaultMessage() throws Exception {
         this.mockMvc.perform(get(HOME_URL).accept(MediaType.APPLICATION_JSON))
             .andExpect(status().isOk())
             .andExpect(content().string(containsString("Hello World")))
