@@ -1,10 +1,5 @@
 /*----------------------------------------------------------------------------*/
 /* Source File:   LEAPYEARCONTROLLERTEST.JAVA                                 */
-/* Description:   REST Api for LeapYear end-points (Tests).                   */
-/* Author:        Carlos Adolfo Ortiz Quirós (COQ)                            */
-/* Date:          Jul.24/2020                                                 */
-/* Last Modified: Jul.24/2020                                                 */
-/* Version:       1.1                                                         */
 /* Copyright (c), 2020 CSoftZ                                                 */
 /*----------------------------------------------------------------------------*/
 /*-----------------------------------------------------------------------------
@@ -35,13 +30,13 @@ import static org.springframework.restdocs.operation.preprocess.Preprocessors.pr
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
 import static org.springframework.restdocs.webtestclient.WebTestClientRestDocumentation.document;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
 import org.springframework.boot.test.autoconfigure.web.reactive.WebFluxTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.reactive.server.WebTestClient;
 
 import com.example.devops.api.controller.LeapYearController;
@@ -52,13 +47,13 @@ import com.example.devops.service.intr.LeapYearService;
  * <p><b>NOTE:</b>This controller uses Spring WebFlux, this means it is reactive.</p>
  *
  * @author Carlos Adolfo Ortiz Quirós (COQ)
- * @version 1.1, Jul.24/2020
+ * @version 1.2, Dec.12/2020
  * @since 11 (JDK), Jul.24/2020
  */
-@RunWith(SpringRunner.class)
+@ExtendWith(SpringExtension.class)
 @AutoConfigureRestDocs(uriScheme = "http", uriHost = "api.example.com")
 @WebFluxTest(LeapYearController.class)
-public class LeapYearControllerIntegrationTest {
+class LeapYearControllerIntegrationTest {
     private static final String LEAP_YEAR_BASE_URL = "/api/v1/leap/year";
     private static final String LEAP_YEAR_CHECK_URL = LEAP_YEAR_BASE_URL + "/check?leapYear=2020";
     private static final String LEAP_YEARS_CHECK_URL = LEAP_YEAR_BASE_URL + "/check/years?leapYears=1600,1700,1800,1900,2000,2019,2020";
@@ -76,7 +71,7 @@ public class LeapYearControllerIntegrationTest {
      * Spring WebFlux testing capabilities.
      */
     @Test
-    public void shouldReturnCheckedLeapYear() {
+    void shouldReturnCheckedLeapYear() {
         when(leapYearService.check(LEAP_YEAR_TEST_2020)).thenReturn(LEAP_YEAR_TEST_2020_STATUS);
 
         this.webTestClient.get()
@@ -91,7 +86,7 @@ public class LeapYearControllerIntegrationTest {
     }
 
     @Test
-    public void shouldReturnCheckedLeapYears() {
+    void shouldReturnCheckedLeapYears() {
         when(leapYearService.check(LEAP_YEAR_TEST_1600)).thenReturn(LEAP_YEAR_TEST_1600_STATUS);
         when(leapYearService.check(LEAP_YEAR_TEST_1700)).thenReturn(LEAP_YEAR_TEST_1700_STATUS);
         when(leapYearService.check(LEAP_YEAR_TEST_1800)).thenReturn(LEAP_YEAR_TEST_1800_STATUS);
