@@ -1,10 +1,5 @@
 /*----------------------------------------------------------------------------*/
 /* Source File:   PERSONPRINTERTEST.JAVA                                      */
-/* Description:   Person Printer tests.                                       */
-/* Author:        Carlos Adolfo Ortiz Quirós (COQ)                            */
-/* Date:          Mar.04/2019                                                 */
-/* Last Modified: Jan.27/2020                                                 */
-/* Version:       1.1                                                         */
 /* Copyright (c), 2019, 2020 CSoftZ                                           */
 /*----------------------------------------------------------------------------*/
 /*-----------------------------------------------------------------------------
@@ -15,7 +10,7 @@ package com.example.devops.common.util;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import com.example.devops.domain.Person;
 
@@ -25,29 +20,37 @@ import com.example.devops.domain.Person;
  * Ignoring Lombok Code in Jacoco</a></p>
  *
  * @author Carlos Adolfo Ortiz Quirós (COQ)
- * @version 1.1, Jan.27/2020
+ * @version 1.1, Dec.12/2020
  * @since 11 (JDK), Mar.04/2019
  */
-public class PersonPrinterTest {
+class PersonPrinterTest {
     /**
      * Validates the Person creation holds a complete record.
      */
     @Test
-    public void shouldPersonBeComplete() {
-        Person harrison = Person.builder()
-            .firstName("John").lastName("Harrison").build();
+    void shouldPersonBeComplete() {
+        Person harrison =
+            Person
+                .builder()
+                .firstName("John")
+                .lastName("Harrison")
+                .build();
 
         assertThat(new PersonPrinter(harrison).toString())
-            .isEqualTo("John Harrison");
+            .hasToString("John Harrison");
     }
 
     /**
      * Validates Person Last Name is empty.
      */
     @Test
-    public void shouldPersonLastNameBeEmpty() {
-        Person anonymous = Person.builder()
-            .firstName("anonymous").lastName("").build();
+    void shouldPersonLastNameBeEmpty() {
+        Person anonymous =
+            Person
+                .builder()
+                .firstName("anonymous")
+                .lastName("")
+                .build();
 
         assertThat(new PersonPrinter(anonymous).toString())
             .isEmpty();
