@@ -1,10 +1,5 @@
 /*----------------------------------------------------------------------------*/
 /* Source File:   PERSON.JAVA                                                 */
-/* Description:   Domain object to represent a Person.                        */
-/* Author:        Carlos Adolfo Ortiz Quirós (COQ)                            */
-/* Date:          Mar.01/2019                                                 */
-/* Last Modified: Jul.24/2020                                                 */
-/* Version:       1.2                                                         */
 /* Copyright (c), 2019, 2020 CSoftZ                                           */
 /*----------------------------------------------------------------------------*/
 /*-----------------------------------------------------------------------------
@@ -24,7 +19,7 @@ import lombok.Value;
  * <p>To achieve immutability we use Lombok <a href="https://projectlombok.org/features/Value">@Value</a> annotation.
  *
  * @author Carlos Adolfo Ortiz Quirós (COQ)
- * @version 1.1, Mar.16/2019
+ * @version 1.2, Dec.12/2020
  * @since 11 (JDK), Mar.01/2019
  */
 @Value
